tests/wpt/web-platform-tests/css/cssom-view/MediaQueryList-addListener-removeListener.html

<!DOCTYPE html>
<meta charset="utf-8">
<meta name="flags" content="dom">
<title>CSS Test: CSSOM View MediaQueryList::{add,remove}Listener</title>
<link rel="author" title="Rune Lillesveen" href="mailto:rune@opera.com">
<link rel="help" href="http://www.w3.org/TR/cssom-view/#the-mediaquerylist-interface">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="resources/matchMedia.js"></script>
<div id="log"></div>
<script>
"use strict";

test(function() {
    const mql = window.matchMedia("all");
    assert_inherits(mql, "addListener");
    assert_equals(typeof mql.addListener, "function");
}, "MediaQueryList::addListener is a function");

promise_test(async t => {
    const iframe = await createIFrame(t, 200, 100);
    const heightMQL = iframe.contentWindow.matchMedia("(max-height: 50px)");
    const widthMQL = iframe.contentWindow.matchMedia("(min-width: 150px)");

    let heightEvent;
    let widthEvent;

    heightMQL.addListener(event => {
        heightEvent = event;
    });
    widthMQL.addListener(event => {
        widthEvent = event;
    });

    assert_false(heightMQL.matches);
    assert_true(widthMQL.matches);

    iframe.height = "50"; // 200x100 => 200x50
    await waitForChangesReported();

    assert_equals(heightEvent.media, heightMQL.media);
    assert_true(heightEvent.matches);
    assert_true(heightMQL.matches);
    assert_true(widthMQL.matches);

    iframe.width = "100"; // 200x50 => 100x50
    await waitForChangesReported();

    assert_equals(widthEvent.media, widthMQL.media);
    assert_false(widthEvent.matches);
    assert_false(widthMQL.matches);
    assert_true(heightMQL.matches);
}, "listeners are called when <iframe> is resized");

promise_test(async t => {
    const mql = await createMQL(t);

    let eventsCount = 0;
    mql.addListener(() => {
        eventsCount++;
    });

    for (let i = 1; i <= 10; i++) {
        triggerMQLEvent(mql);
        await waitForChangesReported();
        assert_equals(eventsCount, i);
    }
}, "listeners are called correct number of times");

promise_test(async t => {
    const mql = await createMQL(t);
    const calls = [];

    mql.addListener(() => {
        calls.push("1st");
    });
    mql.addListener(() => {
        calls.push("2nd");
    });

    triggerMQLEvent(mql);
    await waitForChangesReported();

    assert_equals(calls[0], "1st");
    assert_equals(calls[1], "2nd");
}, "listeners are called in order they were added");

promise_test(async t => {
    const mql = await createMQL(t);

    let called = 0;
    const listener = () => {
        called++;
    };

    mql.addListener(listener);
    mql.addListener(listener);

    triggerMQLEvent(mql);
    await waitForChangesReported();

    assert_equals(called, 1);
}, "listener that was added twice is called only once");

test(function() {
    const mql = window.matchMedia("all");
    assert_inherits(mql, "removeListener");
    assert_equals(typeof mql.removeListener, "function");
}, "MediaQueryList::removeListener is a function");

promise_test(async t => {
    const mql = await createMQL(t);
    const listener = t.unreached_func("should not be called");

    mql.addListener(listener);
    mql.removeListener(listener);

    triggerMQLEvent(mql);
    await waitForChangesReported();
}, "MediaQueryList::removeListener removes added listener");
</script>
